% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_land}
\alias{data_land}
\title{Landscape analysis}
\format{
A data frame with 149 observations and 28 columns
\describe{
  \item{coder}{who coded the sign: Factor (3 levels "C","E","S")  (anonymized)}
  \item{researcher}{who took a photograph of the sign: Factor (2 levels "L","S")  (anonymized)}
  \item{sign}{where the sign was found: Factor (11 levels "digi","door","graf",...)}
  \item{type.of.sign}{kind of sign: Factor (5 levels "com","commem","infra","reg","trans")}
  \item{permanent}{was the sign permanent? Factor (2 levels "no","yes")}
  \item{proper.noun}{kind of proper noun on the sign: Factor (9 levels "bn","bn+","cn",...)}
  \item{no.languages}{number of languages on sign: Factor (4 levels "1","2","3","4+")}
  \item{French}{French on sign? Factor (2 levels "0","1")}
  \item{Dutch}{Dutch on sign? Factor (2 levels "0","1")}
  \item{English}{English on sign? Factor (2 levels "0","1")}
  \item{Italian}{Italian on sign? Factor (2 levels "0","1")}
  \item{Spanish}{Spanish on sign? Factor (2 levels "0","1")}
  \item{German}{German on sign? Factor (2 levels "0","1")}
  \item{Indian,Mandarin,Greek,Indonesian,Portuguese,Libanese,Japanese,Russian,Danish,
            Norwegian,Hebrew,Catalan}{12 infrequent languages: Factor (2 levels "0","1")}
  \item{bn.unclear}{brand name unclear: Factor (2 levels "0","1")}
  \item{multilingual.type}{type of multilingualism on sign: Factor (6 levels "0","1","2","3","4","uc")}
  \item{location}{location of sign: Factor (2 levels "M","P")  (anonymized)}
}
}
\source{
Sarah Buschfeld, TU Dortmund
}
\usage{
data_land
}
\description{
The use of language(s) on public signs and categorization criteria.
}
\keyword{datasets}
