% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_speaker}
\alias{data_speaker}
\title{Subject pronouns and a predictor with one very frequent level}
\format{
A data frame with 3370 observations and 6 columns
\describe{
  \item{class}{subject pronoun realized? Factor (2 levels "zero","realized")}
  \item{AGE}{age: Numerical (in months)}
  \item{ETHN_GROUP}{ethnic group: Factor (3 levels "C","I","n_a") (anonymized)}
  \item{MLU}{mean length of utterance: Factor (5 levels "1","2","3","adult","OL")}
  \item{PRN_TYPE}{pronoun type: Factor (5 levels "dem","it_con","it_ex","it_ref","refer")}
  \item{SPEAKER}{speaker: Factor (2 levels "adult","child")}
}
}
\source{
Sarah Buschfeld, TU Dortmund
}
\usage{
data_speaker
}
\description{
Usage of subject pronouns and its predictors; speaker level "adult" very frequent.
}
\keyword{datasets}
