% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_vowel}
\alias{data_vowel}
\title{Vowel length}
\format{
A data frame with 82 observations and 22 columns
\describe{
  \item{Nickname}{nickname: Factor (43 levels "Nick1","Nick2",...,"Nick43") (anonymized)}
  \item{LiBa}{linguistic background: Factor (2 levels "mono","multi")}
  \item{MLU}{mean length of utterance: Factor (3 levels "1","2","3")}
  \item{phone_label}{phone label: Factor (2 levels "fleece","kit")}
  \item{lexeme}{lexeme: Factor (14 levels "bee","cheek","cheese","chicken", ...)}
  \item{phone_left_1_duration}{duration of phone to the left of the vowel: Numerical (in msec)}
  \item{phone_right_1_duration}{duration of phone to the right of the vowel: Numerical (in msec)}
  \item{word_duration}{duration of word: Numerical (in msec)}
  \item{vowel_minimum_pitch}{minimum pitch of vowel: Numerical (in Hertz)}
  \item{vowel_maximum_pitch}{maximum pitch of vowel: Numerical (in Hertz)}
  \item{vowel_intensity_mean}{mean intensity of vowel: Numerical (in decibel)}
  \item{f1_fifty}{first formant F1 at midpoint of vowel (50\%): Numerical (in Hertz)}
  \item{f2_fifty}{second formant F2 at midpoint of vowel (50\%): numercial (in Hertz)}
  \item{target}{vowel length: Numerical (in msec)}
  \item{cons_class_l}{class of consonant to the left of the vowel: Factor (6 levels "l","r","tsh",...)}
  \item{cons_class_r}{class of consonant to the right of the vowel: Factor (7 levels "?"(glottal stop),"empty","nas",...)}
  \item{ETH}{ethnic group: Factor (6 levels "C1a","C1b","C1c","C2a","C2b","C2c") (anonymized)}
  \item{SEX}{gender: Factor (2 levels "female","male)}
  \item{AGE}{age: Numerical (in months)}
  \item{syllables}{number of syllables in lexeme: integer (1,2)}
  \item{speed}{speed of speech: Numerical (word duration / syllables; in msec)}
  \item{country}{country: Factor (2 levels "E","S")  (anonymized)}
}
}
\source{
Sarah Buschfeld, TU Dortmund
}
\usage{
data_vowel
}
\description{
Vowel length and categorization criteria.
}
\keyword{datasets}
