% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_zero}
\alias{data_zero}
\title{Subject pronouns}
\format{
A data frame with 1024 observations and 7 columns
\describe{
  \item{real}{subject pronoun realized? Factor (2 levels "zero","realized")}
  \item{AGE}{age: Numerical (in months)}
  \item{LiBa}{linguistic background: Factor (3 levels "mono","multi", NA)}
  \item{ETH}{ethnic group: Factor (6 levels "C1a","C1b","C1c","C2a","C2b","C2c") (anonymized)}
  \item{SEX}{gender: Factor (2 levels "female","male")}
  \item{MLU}{mean length of utterance: Factor (4 levels "1","2","3","OL")}
  \item{PRN}{pronoun: Factor (7 levels "I","you_s","he","she","it","we","they")}
}
}
\source{
Sarah Buschfeld, TU Dortmund
}
\usage{
data_zero
}
\description{
Usage of subject pronouns and its predictors.
}
\keyword{datasets}
