% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{participant_zero}
\alias{participant_zero}
\title{Participants of subject pronoun study}
\format{
A dataset with 1024 observations and 1 column
\describe{
  \item{participant}{participant of study (P1 - P51)}
}
}
\source{
Sarah Buschfeld, TU Dortmund
}
\usage{
participant_zero
}
\description{
Participants (random excerpt according to data_zero)
}
\keyword{datasets}
