% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrInDTreg.R
\name{PrInDTreg}
\alias{PrInDTreg}
\title{Regression tree resampling by the PrInDT method}
\usage{
PrInDTreg(datain, regname, ctestv=NA, N, pobs, ppre, conf.level=0.95)
}
\arguments{
\item{datain}{Input data frame with class factor variable 'classname' and the\cr
influential variables, which need to be factors or numericals (transform logicals and character variables to factors)}

\item{regname}{name of regressand variable (character)}

\item{ctestv}{Vector of character strings of forbidden split results;\cr
{see function \code{\link{PrInDT}} for details.}\cr
If no restrictions exist, the default = NA is used.}

\item{N}{Number of repetitions (integer > 0)}

\item{pobs}{Vector of resampling percentages of observations (numerical, between 0 and 1)}

\item{ppre}{Vector of resampling percentages of predictor variables (numerical, between 0 and 1)}

\item{conf.level}{(1 - significance level) in function \code{ctree} (numerical between 0 and 1);\cr
default = 0.95}
}
\value{
\describe{
\item{meanint}{Mean number of interpretable trees over the combinations of individual percentages in 'pobs' and 'ppre'}
\item{R2mean}{Mean R2 on test sets}
\item{ctmax}{best resampled regression tree according to R2 on the full data set}
\item{percmax}{Maximum R2 achieved for \%observations}
\item{perfeamax}{Maximum R2 achieved for \%predictors}
\item{maxR2}{best R2 on the full data set for resampled regression trees (for 'ctmax')} 
\item{interpmax}{interpretability of best tree 'ctmax'}
\item{ctmax2}{second best resampled regression tree according to R2 on the full data set}
\item{percmax2}{second best R2 achieved for \%observations}
\item{perfeamax2}{second best R2 achieved for \%features}
\item{max2R2}{second best R2 on the full data set for resampled regression trees (for 'ctmax2')}
\item{interp2max}{interpretability of second-best tree 'ctmax2'}
}
}
\description{
Regression tree optimzation to identify the best interpretable tree; interpretability is checked (see 'ctestv').\cr
The relationship between the target variable 'regname' and all other factor and numerical variables
in the data frame 'datain' is optimally modeled by means of 'N' repetitions of subsampling.\cr 
The optimization criterion is the R2 of the model on the full sample.\cr
Multiple subsampling percentages of observations and predictors can be specified (in 'pobs' and 'ppre', correspondingly).\cr
The trees generated from undersampling can be restricted by
rejecting unacceptable trees which include split results specified in the character strings of the vector 'ctestv'.\cr
}
\details{
For the optimzation of the trees, we employ a method we call Sumping (Subsampling umbrella of 
model parameters), a variant of Bumping (Bootstrap umbrella of model parameters) (Tibshirani 
& Knight, 1999) which use subsampling instead of bootstrapping. The aim of the 
optimization is to identify conditional inference trees with maximum predictive power
on the full sample under interpretability restrictions. 

\strong{Reference}\cr Tibshirani, R., Knight, K. 1999. Model Search and Inference By Bootstrap "bumping".
           Journal of Computational and Graphical Statistics, Vol. 8, No. 4 (Dec., 1999), pp. 671-686 
 
Standard output can be produced by means of \code{print(name)} or just \code{ name } as well as \code{plot(name)} where 'name' is the output data 
frame of the function.\cr
The plot function will produce a series of more than one plot. If you use R, you might want to specify \code{windows(record=TRUE)} before 
\code{plot(name)} to save the whole series of plots. In R-Studio this functionality is provided automatically.
}
\examples{
data <- PrInDT::data_vowel
data <- na.omit(data)
ctestv <- 'vowel_maximum_pitch <= 320'
N <- 30 # no. of repetitions
pobs <- c(0.70,0.60) # percentages of observations
ppre <- c(0.90,0.70) # percentages of predictors
outreg <- PrInDTreg(data,"target",ctestv,N,pobs,ppre)
outreg
plot(outreg)

}
